stack traces and more

+++ kernel/lib/math.c
@@ -1,6 +1,6 @@
 #include "math.h"
 
-i32 abs(i32 val) {
+i64 abs(i64 val) {
     if(val < 0) return val * -1;
     return val;
 }
